requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
alpine==1.0.14
asgiref==3.7.2
attrs==24.2.0
boto3==1.34.61
botocore==1.34.61
certifi==2024.8.30
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
Django==4.2.9
django-cors-headers==4.5.0
django-crispy-forms==2.1
django-storages==1.14.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
frozenlist==1.5.0
gunicorn==21.2.0
idna==3.7
jmespath==1.0.1
multidict==6.1.0
mysqlclient==2.2.4
numpy==2.1.3
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pillow==11.3.0
propcache==0.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.10
pyactiveresource==2.2.2
PyJWT==2.9.0
python-barcode==0.15.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
qrcode==7.4.2
requests==2.31.0
s3transfer==0.10.0
ShopifyAPI==12.7.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.18
whitenoise==6.9.0
yarl==1.17.1
